+++ impl/filter_lib/fo_apf.hpp
@@ -9,5 +9,5 @@
 
 class FO_APF : public Biquad {
 public:
-    tp_coeffs& calculate_coeffs(float filter_frequency_hz, float sampling_rate_hz = 44100);
+    tp_coeffs& calculate_coeffs(double filter_frequency_hz, double sampling_rate_hz = 44100);
 };
